refactor(cli): decompose ink-ui-bridge into surface/live-run/event-log modules
InkUiBridge (1000 lines) splits along its real seams, public API and all
67 bridge specs unchanged:

- ink/tool-input-format.ts + ink/tool-output-format.ts: the pure
  formatting layer (input summaries, shell/path compaction, whole-word
  tool classification; result summaries/previews, output-text
  extraction, error detection).
- ink/event-text.ts: ANSI/fence/surrogate-safe event-text truncation.
- ink/event-log.ts: the append-only transcript store plus the one-beat
  tool-trace merge buffer (·×N) and its flush ordering rule.
- ink/live-run.ts: the in-flight run region — streaming text, live tool
  lines, the abort latch, and finalize semantics (narration vs answer).
- ink/bridge-surface.ts: abstract base carrying the thin message surface
  (info/warn/section/runSummary/clarification/…) over the abstract
  primitives.
- ink/ink-ui-bridge.ts keeps the snapshot + emit throttle and delegates;
  events/liveText/liveTools stay readable on the instance (tests and
  callers relied on the runtime shape).
243/202/250/96/57/191/113 lines.
